Changed file name back to Duke Level 2 done
